Removed functions not used
Removed any functions that were not called. Added new context Locations
to hold the list of locations and their codes.
